Move download utilities to more appropriate place
